ai_scientist/requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-upstage
streamlit